requirements.txt
asgiref==3.3.4
colorama==0.4.4
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
Pillow==8.2.0
pymaging==0.1
pymaging-png==0.1
pytz==2021.1
qrcode==6.1
six==1.15.0
sqlparse==0.4.1
